Move executeContinuously into LegacyWatched
This helps keep Watched.scala more manageable until we can potentially
remove this code.
