+++ main-command/src/main/scala/sbt/Watched.scala
@@ -10,9 +10,15 @@ package sbt
 import java.io.File
 import java.nio.file.FileSystems
 
-import sbt.BasicCommandStrings._
+import sbt.BasicCommandStrings.{
+  ContinuousExecutePrefix,
+  FailureWall,
+  continuousBriefHelp,
+  continuousDetail
+}
 import sbt.BasicCommands.otherCommandParser
 import sbt.CommandUtil.withAttribute
+import sbt.internal.LegacyWatched
 import sbt.internal.io.{ EventMonitor, Source, WatchState }
 import sbt.internal.util.AttributeKey
 import sbt.internal.util.Types.const
@@ -227,48 +233,8 @@ object Watched {
   }
 
   @deprecated("Replaced by Watched.command", "1.3.0")
-  def executeContinuously(watched: Watched, s: State, next: String, repeat: String): State = {
-    @tailrec def shouldTerminate: Boolean =
-      (System.in.available > 0) && (watched.terminateWatch(System.in.read()) || shouldTerminate)
-    val log = s.log
-    val logger = new EventMonitor.Logger {
-      override def debug(msg: => Any): Unit = log.debug(msg.toString)
-    }
-    s get ContinuousEventMonitor match {
-      case None =>
-        // This is the first iteration, so run the task and create a new EventMonitor
-        (ClearOnFailure :: next :: FailureWall :: repeat :: s)
-          .put(
-            ContinuousEventMonitor,
-            EventMonitor(
-              WatchState.empty(watched.watchService(), watched.watchSources(s)),
-              watched.pollInterval,
-              watched.antiEntropy,
-              shouldTerminate,
-              logger
-            )
-          )
-      case Some(eventMonitor) =>
-        printIfDefined(watched watchingMessage eventMonitor.state)
-        val triggered = try eventMonitor.awaitEvent()
-        catch {
-          case e: Exception =>
-            log.error(
-              "Error occurred obtaining files to watch.  Terminating continuous execution..."
-            )
-            s.handleError(e)
-            false
-        }
-        if (triggered) {
-          printIfDefined(watched triggeredMessage eventMonitor.state)
-          ClearOnFailure :: next :: FailureWall :: repeat :: s
-        } else {
-          while (System.in.available() > 0) System.in.read()
-          eventMonitor.close()
-          s.remove(ContinuousEventMonitor)
-        }
-    }
-  }
+  def executeContinuously(watched: Watched, s: State, next: String, repeat: String): State =
+    LegacyWatched.executeContinuously(watched, s, next, repeat)
 
   private[sbt] object NullLogger extends Logger {
     override def trace(t: => Throwable): Unit = {}
